Redesigned Game Control Panel GUI and Timer Object
I have redesigned the user interface for the Game Control Panel. I have
set up the app to get input from the two drop down menus and display the
output in a message text box. I also created an object for the timer
which will manage the variables that make up the countdown timer.