Multithreading Support
I am now running the loop that updates the timer in a new thread. This
is done so that the application may do other things while the timer is
being updated. The score support will most likely also be implemented as
a separate thread. I also restricted the number of lives inputs in the
range from 1-6 based on input from the gun design team.