Created Installer for Testing Version A
I changed the directories for the application for better organization.
No changes to the code of the application itself. I published the
application to create an installer to test this version of the
application (Version A). This version should fulfill the basic
requirements except for live score updating. This version of the app
will behave as if there were no shots fired during the entire game.
Every message box that pops up will represent a message sent to the
Beaglebone through the server's communication. The actual details of the
message communication will be completed later. Please try to install and
use the application on your PC so that I can get feedback at class
tomorrow for changes to make.