Separating score controls into new class
The right half of the user interface has the controls used to view the
game's score. This was originally built as one class, but I have
separated these controls into a new scorekeeping class. This will allow
me to use object oriented programming to encapsulate this part of the
program. It will also improve code reusability by making it easy to
duplicate the score keeping elements in the second window of the app.