File Read/Write and Scoring
I have added support for reading and writing from a score file. There
should be further testing on the results of the scoring algorithm. The
scoring rules should be further documented. I need to add support for
automatically refreshing the score data. I need to archive the score
data when the game is over. I need to test live updating the score file.
I need to make the game end when only one player is still alive. There
is still a lot of work left on the Scoring class, but I am close to
completing this class.