Testing and further documenting ScoreControl
Completed additional testing on the ScoreControl class and added
additional comments to document the code. I also implemented constant
string keywords that the web server will scan for to send messages to
STARTGAME, RESUMEGAME, and ENDGAME. The RESUMEGAME and ENDGAME keywords
are parameterless, but the STARTGAME keyword requires the web server to
also record the digit following the keyword because this contains the
number of lives the players will have in the game.