Added Help File and Version Information
I have completed a compiled Windows Help File and Help Website for the
Game Control Panel. I added a menu bar that can be used to open the help
file and view the about window (version information). Added help topics
for how to view the About Window and troubleshooting if the user tries
to install the app when it is already installed. I also deleted the
backup of the Version 1.0 folder because the Github release feature
makes this redundant.